pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ruinenhaft <eulennests@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
mssql-django = "^1.2"
mysqlclient = "^2.1.1"
pandas = "^1.5.3"
pymysql = "^1.0.2"
sqlalchemy = "^2.0.5.post1"
opencv-python = "^4.7.0.72"
matplotlib = "^3.7.1"
google-cloud-vision = "^3.4.0"
elasticsearch = "7"
mysql = "^0.0.3"
mysql-connector-python = "^8.0.32"
elasticsearch-dsl = "7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
